Add ip filter config
PR adds a default-storage.yml file with storage items that are applied by default.  This allows the ip filter policy to be auto applied to every service.
Also added the ability to set the services key to '__all__' to apply to all services.
Unit tests added for the make-storage command
Also rnamed make-cloudformation to make-storage which seems more logical