Fix "duplicate SSM parameter name" issue in Opensearch addon template
The error was due to having two SSM parameters with the same SSM name, even though they used conditions to ensure only one parameter was enabled at a time.

THe solution was to use an Fn::if statement to select the endpoint url from the HA or non HA opensearch domain based on the EnableHA condition.  However, this resulted in another issue (possibly a CFN bug, or limitation) due to a Fn:If and a complicated nested statement. To solve this I have created an intermediate SSM parameter which stores the domain endpoint url, then a final config parameter which retreives the endpoint url, and the username and pasword from secrets manager.  Overall we have 1 secrets manager secret and two SSM parameters, which is not ideal, but required to overcome cfn limitations.