fix: Watching ECR pushes where branch name contains a /
`ci-image-builder` already automatically converts a `/` in a branch name to a `-`

This functionality is required because the set of allowed characters in a docker tag name is a lot more restricted compared to the set of allowed characters in a git branch name.

Replicate this conversion here, so that users are unaware of the conversion between `/` and `-` behind the scenes.